Move constructors to own files
