+++ packages/ingest-mongodb-public/src/sources/mongoose.ts
@@ -1,0 +1,32 @@
+import { makeAcquitRequireMdOnGithubDataSource } from "mongodb-rag-ingest/sources";
+
+export const mongooseSourceConstructor = async () => {
+  const repoUrl = "https://github.com/Automattic/mongoose";
+  const testFileLoaderOptions = {
+    branch: "master",
+    recursive: true,
+    ignoreFiles: [/^(?!^\/test\/).+$/],
+  };
+  const repoLoaderOptions = {
+    branch: "master",
+    recursive: true,
+    ignoreFiles: [/^(?!^\/docs\/).+$/],
+  };
+  return await makeAcquitRequireMdOnGithubDataSource({
+    repoUrl,
+    repoLoaderOptions,
+    name: "mongoose",
+    pathToPageUrl(path) {
+      return path
+        .replace(/^\/docs\//, "https://mongoosejs.com/docs/")
+        .replace(/\.md$/, ".html");
+    },
+    testFileLoaderOptions,
+    acquitCodeBlockLanguageReplacement: "javascript",
+    metadata: {
+      productName: "Mongoose ODM",
+      tags: ["node.js", "community library", "mongoose", "odm"],
+      version: "v7.x (current)",
+    },
+  });
+};
